chore: refactor database transfer
