+++ frontend/src/components/TransferOutDatabaseForm/TransferOutDatabaseForm.vue
@@ -1,17 +1,29 @@
 <template>
-  <DrawerContent :title="$t('database.transfer-database-from-to')">
+  <DrawerContent :title="$t('database.transfer-database-to')">
     <div
       class="w-[calc(100vw-8rem)] lg:w-[60rem] max-w-[calc(100vw-8rem)] h-full flex flex-col gap-y-2"
     >
-      <div class="flex items-center justify-between gap-x-8">
-        <div class="flex-1 flex items-center gap-x-2">
-          <span class="textlabel">
-            {{ $t("database.transfer.source-project") }}
+      <div
+        v-if="loading"
+        class="absolute inset-0 z-10 bg-white/70 flex items-center justify-center"
+      >
+        <BBSpin />
+      </div>
+      <div v-else class="space-y-4">
+        <div>
+          <span class="text-main">
+            {{ $t("database.transfer.select-databases") }}
+            <span class="text-red-500">*</span>
           </span>
-          <ProjectV1Name :project="sourceProject" :link="false" />
+          <MultipleDatabaseSelector
+            v-model:selected-uid-list="selectedUidList"
+            :transfer-source="'OTHER'"
+            :database-list="[...databaseList]"
+          />
         </div>
-        <div class="flex-1 flex items-center gap-x-2">
-          <span class="textlabel">
+        <NDivider class="w-full py-2" />
+        <div class="space-y-4">
+          <span class="text-main">
             {{ $t("database.transfer.select-target-project") }}
             <span class="text-red-500">*</span>
           </span>
@@ -19,31 +31,8 @@
             v-model:project="targetProjectId"
             :allowed-project-role-list="[PresetRoleType.OWNER]"
             :include-default-project="true"
-            :filter="filterTargetProject"
           />
         </div>
-      </div>
-      <NTransfer
-        ref="transfer"
-        v-model:value="selectedValueList"
-        :options="sourceTransferOptions"
-        :render-source-list="renderSourceList"
-        :render-target-list="renderTargetList"
-        :source-filterable="true"
-        :source-filter-placeholder="$t('common.filter-by-name')"
-        class="bb-transfer-out-database-transfer"
-        style="height: 100%"
-      />
-      <div
-        class="absolute top-1/2 left-1/2 -translate-x-1/2 z-10 w-8 h-8 text-control flex items-center justify-center"
-      >
-        <heroicons:chevron-right class="w-8 h-8" />
-      </div>
-      <div
-        v-if="loading"
-        class="absolute inset-0 z-10 bg-white/70 flex items-center justify-center"
-      >
-        <BBSpin />
       </div>
     </div>
 
@@ -73,36 +62,19 @@
 </template>
 
 <script setup lang="ts">
-import {
-  NTransfer,
-  NTree,
-  TreeOption,
-  TransferRenderSourceList,
-  NButton,
-  NTooltip,
-} from "naive-ui";
-import { computed, h, ref, toRef } from "vue";
+import { NButton, NTooltip, NDivider } from "naive-ui";
+import { computed, ref, watch } from "vue";
 import { useI18n } from "vue-i18n";
-import { ProjectV1Name, ProjectSelect, DrawerContent } from "@/components/v2";
+import { ProjectSelect, DrawerContent } from "@/components/v2";
 import {
   pushNotification,
   useDatabaseV1Store,
-  useProjectV1ByUID,
   useProjectV1Store,
 } from "@/store";
-import { PresetRoleType, UNKNOWN_ID } from "@/types";
-import { Project } from "@/types/proto/v1/project_service";
-import Label from "./Label.vue";
-import {
-  DatabaseTreeOption,
-  flattenTreeOptions,
-  mapTreeOptions,
-  getDatabaseTreeValue,
-  isDatabaseTreeValue,
-} from "./common";
+import { PresetRoleType, ComposedDatabase, UNKNOWN_ID } from "@/types";
 
 const props = defineProps<{
-  projectId: string;
+  databaseList: ComposedDatabase[];
   selectedDatabaseUidList?: string[];
 }>();
 
@@ -114,22 +86,21 @@ const { t } = useI18n();
 const projectStore = useProjectV1Store();
 const databaseStore = useDatabaseV1Store();
 const loading = ref(false);
-const transfer = ref<InstanceType<typeof NTransfer>>();
 
-const databaseList = computed(() => {
-  const project = projectStore.getProjectByUID(props.projectId);
-  return databaseStore.databaseListByProject(project.name);
-});
+const selectedUidList = ref<string[]>(props.selectedDatabaseUidList ?? []);
 
-const selectedValueList = ref<string[]>(
-  (props.selectedDatabaseUidList ?? []).map(getDatabaseTreeValue)
+watch(
+  () => props.selectedDatabaseUidList ?? [],
+  (list) => (selectedUidList.value = list),
+  { immediate: true }
 );
+
 const selectedDatabaseList = computed(() => {
-  return selectedValueList.value.filter(isDatabaseTreeValue).map((value) => {
-    const uid = value.split("-").pop()!;
+  return selectedUidList.value.map((uid) => {
     return databaseStore.getDatabaseByUID(uid);
   });
 });
+
 const targetProjectId = ref<string>();
 const targetProject = computed(() => {
   const id = targetProjectId.value;
@@ -137,14 +108,12 @@ const targetProject = computed(() => {
   return projectStore.getProjectByUID(id);
 });
 
-const { project: sourceProject } = useProjectV1ByUID(toRef(props, "projectId"));
-
 const validationErrors = computed(() => {
   const errors: string[] = [];
   if (!targetProject.value) {
     errors.push(t("database.transfer.errors.select-target-project"));
   }
-  if (selectedDatabaseList.value.length === 0) {
+  if (selectedUidList.value.length === 0) {
     errors.push(t("database.transfer.errors.select-at-least-one-database"));
   }
   return errors;
@@ -154,113 +123,37 @@ const allowTransfer = computed(() => {
   return validationErrors.value.length === 0;
 });
 
-const sourceTreeOptions = computed(() => {
-  return mapTreeOptions(databaseList.value);
-});
-const sourceTransferOptions = computed(() => {
-  return flattenTreeOptions(sourceTreeOptions.value);
-});
-
-const renderSourceList: TransferRenderSourceList = ({ onCheck, pattern }) => {
-  return h(NTree, {
-    style: "margin: 0 4px;",
-    keyField: "value",
-    defaultExpandAll: true,
-    checkable: true,
-    selectable: false,
-    blockLine: true,
-    checkOnClick: true,
-    cascade: true,
-    virtualScroll: true,
-    data: sourceTreeOptions.value,
-    renderLabel: ({ option }: { option: TreeOption }) => {
-      return h(Label, {
-        option: option as DatabaseTreeOption,
-        keyword: pattern,
-      });
-    },
-    pattern,
-    checkedKeys: selectedValueList.value,
-    showIrrelevantNodes: false,
-    onUpdateCheckedKeys: (checkedKeys: string[]) => {
-      onCheck(checkedKeys.filter(isDatabaseTreeValue));
-    },
-  });
-};
-
-const targetTreeOptions = computed(() => {
-  return mapTreeOptions(selectedDatabaseList.value);
-});
-const targetCheckedKeys = computed(() => {
-  return flattenTreeOptions(targetTreeOptions.value).map(
-    (option) => option.value
-  );
-});
-
-const renderTargetList: TransferRenderSourceList = ({ onCheck }) => {
-  return h(NTree, {
-    key: targetCheckedKeys.value.join(","),
-    style: "margin: 0 4px;",
-    keyField: "value",
-    defaultExpandAll: true,
-    selectable: false,
-    blockLine: true,
-    checkable: true,
-    checkOnClick: true,
-    cascade: true,
-    virtualScroll: true,
-    data: targetTreeOptions.value,
-    renderLabel: ({ option }: { option: TreeOption }) => {
-      return h(Label, {
-        option: option as DatabaseTreeOption,
-      });
-    },
-    checkedKeys: targetCheckedKeys.value,
-    onUpdateCheckedKeys: (checkedKeys: string[]) => {
-      onCheck(checkedKeys.filter(isDatabaseTreeValue));
-    },
-  });
-};
-
-const filterTargetProject = (project: Project) => {
-  return project.uid !== props.projectId;
-};
-
 const doTransfer = async () => {
   const target = targetProject.value!;
   if (!target) return;
 
-  const databaseList = selectedDatabaseList.value;
+  const databaseList = selectedDatabaseList.value.filter(
+    (db) => db.project !== target.name
+  );
 
   try {
     loading.value = true;
-    await useDatabaseV1Store().transferDatabases(
-      selectedDatabaseList.value,
-      target.name
-    );
-    const displayDatabaseName =
-      databaseList.length > 1
-        ? `${databaseList.length} databases`
-        : `'${databaseList[0].databaseName}'`;
 
-    pushNotification({
-      module: "bytebase",
-      style: "SUCCESS",
-      title: `Successfully transferred ${displayDatabaseName} to project '${target.title}'.`,
-    });
+    if (databaseList.length > 0) {
+      await useDatabaseV1Store().transferDatabases(
+        selectedDatabaseList.value,
+        target.name
+      );
+      const displayDatabaseName =
+        databaseList.length > 1
+          ? `${databaseList.length} databases`
+          : `'${databaseList[0].databaseName}'`;
+
+      pushNotification({
+        module: "bytebase",
+        style: "SUCCESS",
+        title: `Successfully transferred ${displayDatabaseName} to project '${target.title}'.`,
+      });
+    }
+
     emit("dismiss");
   } finally {
     loading.value = false;
   }
 };
 </script>
-
-<style>
-.bb-transfer-out-database-transfer {
-  @apply gap-x-8;
-}
-
-.bb-transfer-out-database-transfer .n-transfer-list--target {
-  border-left: 1px solid var(--n-border-color) !important;
-}
-</style>
