+++ frontend/src/locales/en-US.json
@@ -792,7 +792,7 @@
           "self": "MD5 mask",
           "salt": "Salt",
           "salt-label": "Salt is used to generate a hash from the original value."
-        } 
+        }
       },
       "masking-level": {
         "self": "Masking Level",
@@ -1710,8 +1710,10 @@
         "select-target-project": "Select target project",
         "select-at-least-one-database": "Select at least one database"
       },
+      "select-databases": "Select databases",
       "select-target-project": "Select target project"
     },
+    "transfer-database-to": "Transfer databases to target project",
     "transfer-database-from-to": "Transfer database from source project to target project",
     "classification": {
       "self": "Classification"
